v0.42 — Room speed-up, broadcasts, self-destruct alerts & QoL
- LiveKit token pre-warming for faster room joins
- Soft update dialog on app launch (reuses existing Firestore read)
- Broadcast banners show gift quantity (3x Crown format)
- Cross-room broadcast banners restored via Firestore observation
- Super Shy purchase stays on sheet instead of blank loading screen
- Self-destruct TTS countdown (opt-in via Settings > Room Alerts)
- ShyBeans redemption supports large Long balances
- Broadcast banner repositioned above seat grid
- Buy coins sheet available inside rooms
- Invite shows error if person has left the room
- Cloud Functions: quantity field in lastGiftEvent for all gift sends

+++ shared/src/commonMain/kotlin/com/shyden/shytalk/data/repository/EconomyRepository.kt
@@ -17,7 +17,7 @@ interface EconomyRepository {
     suspend fun sendGiftDirect(recipientId: String, giftId: String, quantity: Int = 1): Resource<Map<String, Any?>>
     suspend fun sendGiftBatch(recipientIds: List<String>, giftId: String, quantity: Int, fromBackpack: Boolean): Resource<Map<String, Any?>>
     suspend fun sendEntireBackpack(recipientId: String): Resource<Map<String, Any?>>
-    suspend fun redeemBeans(amount: Int): Resource<Map<String, Any?>>
+    suspend fun redeemBeans(amount: Long): Resource<Map<String, Any?>>
     suspend fun purchaseCoins(productId: String, purchaseToken: String): Resource<Map<String, Any?>>
     suspend fun purchaseSubscription(productId: String, purchaseToken: String): Resource<Map<String, Any?>>
     suspend fun getCoinPackages(): Resource<List<CoinPackage>>
